refactor(ThreeVideoDistortion): sperate components by file
